Initial commit
Created three files. Item.py contains the Item class for items that will be available in the vending machine, Vendor.py contains the Vendor class that stores the variables and functions to add and purchase items from the vending machine, and Project1.py will contain the runtime loop to instantiate Vendor and allow user input.